+++ packages/themes/Fundu/src/Theme/FunduTheme.js
@@ -11,6 +11,7 @@ import {
   ExpandListItemStyling
 } from "./theme.Lists"
 import { TextStyling } from "./theme.Texts"
+import { ImageStyling } from "./theme.Images"
 import { ButtonStyling } from "./theme.Buttons"
 import {
   InputStyling,
@@ -60,10 +61,11 @@ const spacing = {
   small: "12px",
   tiny: "6px"
 }
+
 const color = {
-  bg: "#fcfcfc",
+  bg: "#F5F8FC",
   bgGray: "#F9F7F7",
-  primary: "#23A532",
+  primary: "#1F5D9B",
   gray: "#767676",
   text: "#2b2b2b",
   white: "#fff",
@@ -87,27 +89,20 @@ const font = {
   h4: "18px",
   h5: "16px",
   h6: "14px",
-  base: "14px",
+  base: "16px",
   tiny: "12px",
   input: "16px",
-  huge: "44px"
-}
-
-const webfonts = {
-  google: {
-    families: ["Libre Baskerville:700", "Libre Franklin:400,600"]
-  }
+  huge: "50px"
 }
 
 const fontFamily = {
-  heading: "Libre Baskerville",
-  body: "Libre Franklin"
+  heading: "AvenirLTPro-Black",
+  body: "AvenirLTPro-Roman"
 }
 
 const theme = new theming.Theme(BaseTheme, {
   name: "Fundu",
   font,
-  webfonts,
   color,
   spacing,
   wrapper,
@@ -147,4 +142,5 @@ const theme = new theming.Theme(BaseTheme, {
   .add(LogoStyle)
   .add(SelectSelectedWrapperStyling)
   .add(PostalCodeInputStyling)
+  .add(ImageStyling)
 export default theme