+++ packages/themes/Fundu/src/Theme/theme.Images.js
@@ -1,0 +1,11 @@
+import { css } from "styled-components"
+import { Image, theming } from "@staccx/base"
+
+export const ImageStyling = theming.createVariants(
+  {
+    industryCard: css`
+      max-height: 100%;
+    `
+  },
+  Image.themeProps.image
+)