+++ packages/apps/bank-in-a-box/CHANGELOG.md
@@ -3,6 +3,23 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="7.16.0"></a>
+# [7.16.0](https://bitbucket.org/stacc-flow/bento/compare/v7.15.0...v7.16.0) (2018-10-25)
+
+
+### Bug Fixes
+
+* **bank-in-a-box:** fix some bugs on transactions ([ad9802b](https://bitbucket.org/stacc-flow/bento/commits/ad9802b))
+
+
+### Features
+
+* **bank-in-a-box:** autofill to-account now works ([cacb53c](https://bitbucket.org/stacc-flow/bento/commits/cacb53c))
+
+
+
+
+
 <a name="7.15.0"></a>
 # [7.15.0](https://bitbucket.org/stacc-flow/bento/compare/v7.14.0...v7.15.0) (2018-10-22)